backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Werkzeug==3.0.1
PyJWT==2.8.0
python-dotenv>=1.0.0
requests>=2.31.0
psycopg2-binary>=2.9.10
redis>=5.0.0
pika>=1.3.2
waitress>=3.0.0
dashscope>=1.20.0
imageio-ffmpeg>=0.4.9
eventlet>=0.33.0
websocket-client>=1.6.0
python-socketio>=5.10.0


